requirements.txt
torchvision==0.6.0
tqdm==4.36.1
numpy==1.18.2
torch==1.5.0
Pillow==7.2.0
